perf: float32 数据缓存与减少同步

+++ data_provider/data_loader.py
@@ -95,7 +95,8 @@ class Dataset_ETT_hour(Dataset):
             prior_scale=getattr(self.args, "prior_scale", None),
             prior_offset=getattr(self.args, "prior_offset", None),
         )
-        data = self.scaler.transform(df_data.values)
+        # sklearn 的 StandardScaler 默认输出 float64；这里统一落为 float32，减少 DataLoader 到 GPU 的 dtype 转换开销
+        data = self.scaler.transform(df_data.values).astype(np.float32, copy=False)
 
         df_stamp = df_raw[['date']][border1:border2]
         df_stamp['date'] = pd.to_datetime(df_stamp.date)
@@ -104,10 +105,10 @@ class Dataset_ETT_hour(Dataset):
             df_stamp['day'] = df_stamp.date.apply(lambda row: row.day, 1)
             df_stamp['weekday'] = df_stamp.date.apply(lambda row: row.weekday(), 1)
             df_stamp['hour'] = df_stamp.date.apply(lambda row: row.hour, 1)
-            data_stamp = df_stamp.drop(['date'], 1).values
+            data_stamp = df_stamp.drop(['date'], 1).values.astype(np.float32, copy=False)
         elif self.timeenc == 1:
             data_stamp = time_features(pd.to_datetime(df_stamp['date'].values), freq=self.freq)
-            data_stamp = data_stamp.transpose(1, 0) 
+            data_stamp = data_stamp.transpose(1, 0).astype(np.float32, copy=False)
 
         self.data_x = data[border1:border2]
         self.data_y = data[border1:border2]
@@ -195,7 +196,7 @@ class Dataset_ETT_minute(Dataset):
             prior_scale=getattr(self.args, "prior_scale", None),
             prior_offset=getattr(self.args, "prior_offset", None),
         )
-        data = self.scaler.transform(df_data.values)
+        data = self.scaler.transform(df_data.values).astype(np.float32, copy=False)
 
         df_stamp = df_raw[['date']][border1:border2]
         df_stamp['date'] = pd.to_datetime(df_stamp.date)
@@ -206,10 +207,10 @@ class Dataset_ETT_minute(Dataset):
             df_stamp['hour'] = df_stamp.date.apply(lambda row: row.hour, 1)
             df_stamp['minute'] = df_stamp.date.apply(lambda row: row.minute, 1)
             df_stamp['minute'] = df_stamp.minute.map(lambda x: x // 15)
-            data_stamp = df_stamp.drop(['date'], 1).values
+            data_stamp = df_stamp.drop(['date'], 1).values.astype(np.float32, copy=False)
         elif self.timeenc == 1:
             data_stamp = time_features(pd.to_datetime(df_stamp['date'].values), freq=self.freq)
-            data_stamp = data_stamp.transpose(1, 0)
+            data_stamp = data_stamp.transpose(1, 0).astype(np.float32, copy=False)
 
         self.data_x = data[border1:border2]
         self.data_y = data[border1:border2]
@@ -307,7 +308,7 @@ class Dataset_Custom(Dataset):
             prior_scale=getattr(self.args, "prior_scale", None),
             prior_offset=getattr(self.args, "prior_offset", None),
         )
-        data = self.scaler.transform(df_data.values)
+        data = self.scaler.transform(df_data.values).astype(np.float32, copy=False)
 
         df_stamp = df_raw[['date']][border1:border2]
         df_stamp['date'] = pd.to_datetime(df_stamp.date)
@@ -316,10 +317,10 @@ class Dataset_Custom(Dataset):
             df_stamp['day'] = df_stamp.date.apply(lambda row: row.day, 1)
             df_stamp['weekday'] = df_stamp.date.apply(lambda row: row.weekday(), 1)
             df_stamp['hour'] = df_stamp.date.apply(lambda row: row.hour, 1)
-            data_stamp = df_stamp.drop(['date'], 1).values
+            data_stamp = df_stamp.drop(['date'], 1).values.astype(np.float32, copy=False)
         elif self.timeenc == 1:
             data_stamp = time_features(pd.to_datetime(df_stamp['date'].values), freq=self.freq)
-            data_stamp = data_stamp.transpose(1, 0)
+            data_stamp = data_stamp.transpose(1, 0).astype(np.float32, copy=False)
 
         self.data_x = data[border1:border2]
         self.data_y = data[border1:border2]
